// frontend/src/entities/environment/types/mutations.ts
export enum MutationTypes {
    SET_RULER = "SET_RULER",
    SET_GRIDS = "SET_GRIDS",
    SET_NUM_GRIDS = "SET_NUM_GRIDS",
    SET_COLOR_GRIDS = "SET_COLOR_GRIDS",
    SET_GAP_GRIDS = "SET_GAP_GRIDS",
    SET_WIDTH_GRIDS = "SET_WIDTH_GRIDS",
    SET_CARD_IS_DRAG = "SET_CARD_IS_DRAG",
    SET_COLUMN_IS_DRAG = "SET_COLUMN_IS_DRAG",
    SET_ITEM_IS_DRAG = "SET_ITEM_IS_DRAG",
    SET_EDITABLE = "SET_EDITABLE",
    ADD_OPENED_REDACTOR = "ADD_OPENED_REDACTOR",
    REMOVE_OPENED_REDACTOR = "REMOVE_OPENED_REDACTOR",
    SET_ENVIRONMENT_VIEWPORT = "SET_ENVIRONMENT_VIEWPORT",
    SET_REDEFINE_FONTS = "SET_REDEFINE_FONTS",
    SET_FONT_PAGE = "SET_FONT_PAGE",
    SET_SHOW_WINDOW_PICKER = "SET_SHOW_WINDOW_PICKER",
    SET_CODE_HEAD_PAGE = "SET_CODE_HEAD_PAGE",
    UPDATE_CODE_HEAD_PAGE = "UPDATE_CODE_HEAD_PAGE",
    DELETE_CODE_HEAD_PAGE = "DELETE_CODE_HEAD_PAGE",

    UPDATE_ARRAY_CODE_BODY_PAGE = "UPDATE_ARRAY_CODE_BODY_PAGE",
    UPDATE_ARRAY_CODE_HEAD_PAGE = "UPDATE_ARRAY_CODE_HEAD_PAGE",

    SET_CODE_BODY_PAGE = "SET_CODE_BODY_PAGE",
    UPDATE_CODE_BODY_PAGE = "UPDATE_CODE_BODY_PAGE",
    DELETE_CODE_BODY_PAGE = "DELETE_CODE_BODY_PAGE",
}